Report file present changes in files, nr of files counter for verification working

+++ main.py
@@ -39,22 +39,17 @@ def getFileInfo(folder, cursor):
             groupIdentity = getpwuid(st.st_gid).pw_name
             lastModify = st.st_mtime
 
-            #print("Filepath {}".format(filepath))
-            #print("Filesize {}".format(fileSize))
-            #print("OwnerIdentidy {}".format(userIdentiy))
-            #print("GroupIdentify {}".format(groupIdentity))
-            #print("AccessRight {}".format(acessRight))
-            #print("lastModify {}".format(lastModify))
             # Should calculate some hashing
             md5Hash = md5(filepath)
 
             cursor.execute("INSERT INTO info VALUES(?,?,?,?,?,?,?,0)",(filepath,fileSize,userIdentiy,groupIdentity,acessRight,lastModify,md5Hash))
-    cursor.commit()
+
             #compare(cursor, filepath, lastModify, fileSize, userIdentiy, groupIdentity, acessRight)
-    #return {'nrOfDirs':nrOfDirs, 'nrOfFiles':nrOfFiles }
-    #yield nrOfDirs # first time the function is called it return dirs, the function stops here
-    #yield nrOfFiles # Second time the function is called it return nrOfFiles
+
+
+    cursor.commit()
     return (nrOfDirs, nrOfFiles)
+
 
 def md5(fileName):
     md5 = hashlib.md5() # Create md5 object
@@ -71,35 +66,6 @@ def getOldfileInfo(cursor,filepath):
     cursor = cursor.execute('SELECT * FROM info WHERE filepath=?',(filepath,))
     for row in cursor:
         return row
-
-
-def compare(cursor, filepath, lastModify, fileSize, userIdentiy, groupIdentity, acessRight):
-    errorMsg = "*********CHANGED: File {}***********\n".format(filepath)
-    oldInfo = getOldfileInfo(cursor,filepath)
-    if oldInfo == None:
-        # There is no reccord of this file in db
-        print("NEW FILE found: {}".format(filepath))
-    else:
-        # update that the file is going to be checked
-        cursor.execute('UPDATE info SET checked=? WHERE filepath = ?',(1,filepath))
-        ursor.commit()
-        # file exists in db
-        print("File is in db") # debugger
-        if oldInfo[5] != lastModify:
-            errorMsg += "prev changes where made {} new changes {}\n".format(oldInfo[5], lastModify)
-            # File has been modified from db version
-            if oldInfo[1] != int(fileSize):
-                errorMsg += "changed fileSize from {} to {}\n".format(oldInfo[1],fileSize)
-            if oldInfo[2] != userIdentiy:
-                errorMsg += "changed useridentify from {} to {}\n".format(oldInfo[2], userIdentiy)
-            if oldInfo[3] != groupIdentity:
-                errorMsg += "changed groupidentiy from {} to {}\n".format(oldInfo[3], groupIdentity)
-            if oldInfo[4] != str(acessRight):
-                errorMsg += "changed accessright from {} to {}\n".format(oldInfo[4], acessRight)
-
-            print(errorMsg)
-        #else:
-            # File has not been changed since befor
 
 
 def initializationReport(monitoreDirectory, pathVerification, nrOfDir, nrofFiles, startTime, reportFile):
@@ -122,31 +88,35 @@ def initializationMode(args):
 
             print("Verification and report ok")
             # check if verification and report file exists
-            if os.path.isfile(args.V) or os.path.isfile(args.R):
-                # User must do a choice
-                ans = ""
-                while(ans != "yes" and ans != "no"):
-                    ans = input("Should we overwrite verification {} and report {} yes/no : ".format(args.V, args.R))
-                if ans == "no":
-                    print("The files will be preserved, goodbye")
-                    quit() # terminate the program
+            if args.H == "MD-5" or args.H == "SHA-1":
+                if os.path.isfile(args.V) or os.path.isfile(args.R):
+                    # User must do a choice
+                    ans = ""
+                    while(ans != "yes" and ans != "no"):
+                        ans = input("Should we overwrite verification {} and report {} yes/no : ".format(args.V, args.R))
+                    if ans == "no":
+                        print("The files will be preserved, goodbye")
+                        quit() # terminate the program
+                    else:
+                        if os.path.isfile(args.V):
+                            os.remove(args.V)
+                        if os.path.isfile(args.R):
+                            os.remove(args.R)
+                # Continue if this was the users will
+                print("Creates new report file and verification file")
+                startTime = time.time()
+
+                if args.H == "MD-5": # Determine the name of the field in db
+                    cursor = dbCreateTable(args.V,"md5")
                 else:
-                    if os.path.isfile(args.V):
-                        os.remove(args.V)
-                    if os.path.isfile(args.R):
-                        os.remove(args.R)
-            # Continue if this was the users will
-            print("Will create new files")
-            startTime = time.time()
+                    cursor = dbCreateTable(args.V,"sha1")
 
-            if args.H == "MD-5": # Determine the name of the field in db
-                cursor = dbCreateTable(args.V,"md5")
+                nrOfDirs, nrOfFiles = getFileInfo(args.D, cursor)
+                cursor.close() # close db connection
+                initializationReport(args.D, args.V, nrOfDirs, nrOfFiles, startTime, args.R)
+                print("Done with Initialization")
             else:
-                cursor = dbCreateTable(args.V,"sha1")
-
-            nrOfDirs, nrOfFiles = getFileInfo(args.D, cursor)
-            cursor.close() # close db connection
-            initializationReport(args.D, args.V, nrOfDirs, nrOfFiles, startTime, args.R)
+                print("You must choose a hash function either, -H SHA-1 or MD5")
         else:
             print("Verification: {}\n Report: {} \n can't be inside: directory {}\n please specify outside {}".format(args.V, args.R, args.D, args.D))
     else: #isdir args.D
@@ -162,21 +132,95 @@ def verificationMode(args):
             ###########
             # Start verification process
             ##########
-            print("Start verification process")
-            mDirectory, verDB = parseReportFile(args)
+            startTime = time.time()
+            cursor = connect_db(args.V)
+            nrOfWarnings, nrOfDirs, nrOfFiles, ssChangedFiles= compare(args.D, cursor)
+            nrOfWarnings, ssChangedFiles = deletedFiles(cursor, nrOfWarnings, ssChangedFiles)
+
+            reportFileVerification(args.D, args.V, args.R, nrOfDirs, nrOfFiles, nrOfWarnings,startTime, ssChangedFiles)
+            # Clean up
+            cursor.execute('UPDATE info SET checked=? WHERE checked =?',(0,1)) # Changed it back
+            cursor.commit() # Change it back
 
     else: # checkUserInputIfValid
         quit()
 
-def parseReportFile(args):
-    reportInformation = []
-    print("Begin parsing")
-    with open(args.R, newline='') as csvfile:
-        parser = csv.reader(csvfile, delimiter=':')
-        for row in parser:
-            reportInformation.append(row[1])
-    # This information should be enought for now
-    return (reportInformation[0], reportInformation[1])
+def compare(folder, cursor):
+    fileChanged = False
+    nrOfDirs = 0
+    nrOfFiles = 0
+    nrOfWarnings = 0
+    ssChangedFiles = ""
+    for root, dirs, files in os.walk(os.path.abspath(folder), topdown=True):
+        nrOfDirs += 1
+        for name in files:
+            nrOfFiles += 1
+            fileChanged = False
+            filepath = os.path.join(root, name)
+            oldInfo = getOldfileInfo(cursor,filepath)
+            if oldInfo == None:
+                nrOfWarnings +=1
+                print("NEW FILE: {}".format(filepath))
+                ssChangedFiles += "NEW FILE: {}\n".format(filepath)
+            else:
+                # Retrive latest information about files
+                st = os.stat(filepath)
+                acessRight = oct(stat.S_IMODE(st.st_mode)) #wwww.stomp.colorado.edu
+                fileSize = st.st_size
+                userIdentiy = getpwuid(st.st_uid).pw_name
+                groupIdentity = getpwuid(st.st_gid).pw_name
+                lastModify = st.st_mtime
+
+                # Should calculate some hashing
+                md5Hash = md5(filepath)
+
+                #Compare the files with the one in db
+                errorMsg = "CHANGED: File {} ".format(filepath)
+                if oldInfo[5] != lastModify:
+                    fileChanged = True
+                    errorMsg += ", prev changes where made {} new changes {}".format(oldInfo[5], lastModify)
+                    # File has been modified from db version
+                if oldInfo[1] != int(fileSize):
+                    fileChanged = True
+                    errorMsg += ", fileSize from {} to {}".format(oldInfo[1],fileSize)
+                if oldInfo[2] != userIdentiy:
+                    fileChanged = True
+                    errorMsg += ", useridentify from {} to {}".format(oldInfo[2], userIdentiy)
+                if oldInfo[3] != groupIdentity:
+                    fileChanged = True
+                    errorMsg += ", groupidentiy from {} to {}".format(oldInfo[3], groupIdentity)
+                if oldInfo[4] != str(acessRight):
+                    fileChanged = True
+                    errorMsg += ", accessright from {} to {}".format(oldInfo[4], acessRight)
+
+                if fileChanged:
+                    nrOfWarnings +=1
+                    print(errorMsg + "\n")
+                    ssChangedFiles += errorMsg + "\n"
+
+                cursor.execute('UPDATE info SET checked=? WHERE filepath =?',(1,filepath))
+        cursor.commit()
+
+    return (nrOfWarnings,nrOfDirs,nrOfFiles,ssChangedFiles)
+
+def deletedFiles(cursor, nrOfWarnings, ssChangedFiles):
+    cursor = cursor.execute('SELECT * FROM info WHERE checked=?',(0,))
+    for row in cursor:
+        nrOfWarnings += 1
+        print("File been deleted: {}".format(row))
+        ssChangedFiles += "File been deleted: {}\n".format(row)
+
+    return (nrOfWarnings, ssChangedFiles)
+
+def reportFileVerification(monitoreDirectory, pathVerification, reportFile, nrOfDir, nrOfFiles, nrOfWarnings, startTime, ssChangedFiles):
+    ss = "Monitored directory: " + monitoreDirectory + "\nVerification file: " + pathVerification + "\nReport file: "+ reportFile + "\nNr of directorys: " + str(nrOfDir) + "\nNr of files: " + str(nrOfFiles) + "\nNr of warnings: " + str(nrOfWarnings)
+    fprint = open(reportFile,"w")
+    elapsedTime = time.time() - startTime
+    ss += "\nTime to complete in seconds: " + str(elapsedTime) + "\n"
+    ss += ssChangedFiles
+    fprint.write(ss)
+    fprint.close()
+
 
 def removeFiles(args):
     if os.path.isfile(args.V) and os.path.isfile(args.R):
@@ -199,27 +243,11 @@ def checkUserInputIfValid(args):
     if(os.path.isdir(args.D)): # Check if directory exist
         #print("{} exists".format(args.D))
         if args.D not in args.V and args.D not in args.R: # Check if the paths is outside the directory
-            if os.path.isdir(args.V) or os.path.isdir(args.R): # Check that the paths leads to folders
-                print("Need to specify a file for verification file and report file")
+            if os.path.isdir(args.V) or os.path.isdir(args.R): # Check that the paths not leads to folders
+                print("Need to specify a file for verification and report")
             else:
                 print("Verification and report ok")
                 flag = True
-                # check if verification and report file exists
-                #if os.path.isfile(args.V) or os.path.isfile(args.R):
-                    # User must do a choice
-                #    if userChoiceDeleteVerandReport == "no":
-                #        print("The files will be preserved, goodbye")
-                #        quit() # terminate the program
-                #    else:
-                #        removeFiles(args)
-                # Continue if this was the users will
-                #print("Will create new files")
-                #startTime = time.time()
-
-                #if args.H == "MD-5": # Determine the name of the field in db
-                #    cursor = dbCreateTable(args.V,"md5")
-                #else:
-                #    cursor = dbCreateTable(args.V,"sha1")
         else:
             print("Verification: {}\n Report: {} \n can't be inside: directory {}\n please specify outside {}".format(args.V, args.R, args.D, args.D))
     else: #isdir args.D
